Move client to subdir
The client is only intended for use with integration tests.  It should
not be in the same source code levels as the actual proxy code.
